make failing test case to deal with weird overwriting issue
